# docs/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MatchListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchModel"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "pagination"
        ],
        "title": "MatchListResponse",
        "type": "object"
      },
...
      "RefreshRequest": {
        "description": "Optional season selector for a refresh; defaults to latest completed.",
        "properties": {
          "current": {
            "default": false,
            "description": "Refresh the active season instead of the latest",
            "title": "Current",
            "type": "boolean"
          },
          "season_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact fantasy season id",
            "title": "Season Id"
          },
          "season_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact stat season name, e.g. 2025/2026",
            "title": "Season Name"
          }
        },
        "title": "RefreshRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read API and squad optimizer for the Fantasy RPL pipeline, plus the manual ingestion control plane. Read endpoints never call Sports.ru.",
    "title": "Fantasy Analytics API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/ingestion/rpl/refresh": {
      "post": {
        "operationId": "refresh_rpl_admin_ingestion_rpl_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Rpl",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/ingestion/runs/{job_id}": {
      "get": {
        "operationId": "get_run_admin_ingestion_runs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Run Admin Ingestion Runs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/matches": {
      "get": {
        "operationId": "list_matches_matches_get",
        "parameters": [
          {
            "in": "query",
            "name": "season_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season Id"
            }
          },
          {
            "in": "query",
            "name": "tour_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tour Id"
            }
          },
          {
            "in": "query",
            "name": "club_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Club Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Matches",
        "tags": [
          "catalog"
        ]
      }
    },
    "/matches/{match_id}": {
      "get": {
        "operationId": "get_match_matches__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Match",
        "tags": [
          "catalog"
        ]
      }
    },
    "/optimizer/squad": {
      "post": {
        "operationId": "optimize_squad_optimizer_squad_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SquadRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize Squad",
        "tags": [
          "optimizer"
        ]
      }
    },
    "/optimizer/transfers": {
      "post": {
        "operationId": "optimize_transfers_optimizer_transfers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransfersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize Transfers",
        "tags": [
          "optimizer"
        ]
      }
    },
...
    "/players/{player_season_id}": {
      "get": {
        "operationId": "get_player_players__player_season_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_season_id",
            "required": true,
            "schema": {
              "title": "Player Season Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tour_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tour Id"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "default": "poisson_events",
              "enum": [
                "poisson_events",
                "season_mean",
                "recent_form"
              ],
              "title": "Model",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDetailModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Player",
        "tags": [
          "players"
        ]
      }
    },
    "/seasons": {
      "get": {
        "operationId": "list_seasons_seasons_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Seasons",
        "tags": [
          "catalog"
        ]
      }
    },
    "/seasons/{season_id}": {
      "get": {
        "operationId": "get_season_seasons__season_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "season_id",
            "required": true,
            "schema": {
              "title": "Season Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonDetailModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Season",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/tours/{tour_id}": {
      "get": {
        "operationId": "get_tour_tours__tour_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tour_id",
            "required": true,
            "schema": {
              "title": "Tour Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tour",
        "tags": [
          "catalog"
        ]
      }
    }
  }
}